QDoc: Remove Node::setOutputFileName and Node::outputFileName
The `Node` class exposed a getter and a setter for an "outputFileName",
which seems to represent the filename that a `Node` that generates a
file in the output directory would have for that file.

The setter was called in `Generator::beginSubPage`, where the
documentation file was created during the generation phase, but the
value that was set in the node was never accessed in other parts of the
codebase.

It is unclear what the exact purpose of the property was, supposedly
some caching mechanism or a way to access a piece of data that was not
easily accessible from other parts of the codebase.

As the property was unused, it is now removed as dead code.
Hence, all references to `Node::setOutputFileName`,
`Node::outputFileName` and `m_ouputFileName`, the internal storage that
`PageNodes` used for the property, are removed, together with their
documentation, when present.

`Generator::beginSubPage` delegated most of its work to
`Generator::beginFilePage`, which took care of the actual file-creation
logic, and differed from its agent only in that it stored the filename
of the created file into the node that the file was created from.

With `Node::setOutputFileName` being removed, `Generator::beginSubPage`
is simply an alias to `Generator::beginFilePage` such that having both
of them is not useful anymore.

Hence, `Generator::beginSubPage` was removed in favor of
`Generator::beginFilePage`.

`Generator::beginFilePage` itself delegates much of the file-creation
logic to `Generator::openSubPageFile` and, furthermore, more usages of
`beginSubPage` and its associated method `endSubPage` were present in
the documentation compared to the usages of `beginFilePage` and its
associated method `endFilePage`.

Hence, `Generator::beginFilePage` was renamed to
`Generator::beginSubPage` and all usages references to `beginFilePage`
were substituted with references to `Generator::beginSubPage`, such as
to increase the consistency with the naming of
`Generator::openSubPageFile` and to reduce the surface of the diff for
the patch.

`Generator::beginFilePage` had an associated method,
`Generator::endFilePage`, an alias for `Generator::endSubPage`, that is
now removed in favor of `Generator::endSubPage` as a consequence of the
renaming of `Generator::beginFilePage`.

A reference to the lack of storage of the created file name in the node
was removed from the documentation of `Generator::beginFilePage`, now
`Generator::beginSubPage`, as it was not relevant anymore.

Change-Id: I6c9a6ef2db3bf4e1d176c6a2a134ba7dcae3cda5
Reviewed-by: Paul Wicking <paul.wicking@qt.io>