Synchronize generateDetailedQmlMember with HTML
The code more closely resembles that of
HtmlGenerator::generateDetailedQmlMember. The change fixes a bug where
checks were made on the node with shared comments (scn/node) instead
of the individual nodes (sharedNode in the loop).

Also, remove the special case of shared comments that generated no
sections: this allows to remove the generateEndSection variable and to
have more consistent documents.

Change-Id: Ia856bb6afb24bb8973e7a9f314264153f6f8076a
Reviewed-by: Paul Wicking <paul.wicking@qt.io>
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>