Build QtDesigner plugins in all configurations
The QtDesigner plugins were marked as 'tool_plugin', which restricts
their build to release-only, even in debug_and_release builds.  This was
done to avoid building debug plugins for tools that are only build as
release.  However, the designer plugins are also loaded when using
QUiLoader and thus are not exclusively meant for the designer tool.

Task-number: QTBUG-104755
Change-Id: I82404d964b468ada803e575bf24d79a5f8922c1f
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>