QDoc: Avoid losing output for unlinkable AutoLinks in WebXMLGenerator
QDoc recognizes a series of words that it finds in documentation blocks
as "AutoLink" candidates, as per the rules in
"DocParser::isAutoLinkString".

When an "AutoLink" candidate is produced, it will later, during the
generation phase, tentatively be resolved as linking to an element of
the documentation.

When this tentative linking succeeds, generally, it will generate a
link to the found element in the output documentation, falling back to
outputting the word if that is not the case.

In the case of `WebXMLGenerator`, the code structure that takes care of
resolving an "AutoLink" output, in `addAtomElements`, had a structural
defect, such that it would completely erase the "AutoLink" candidate
from the output if an element could be associated with the "AutoLink"
without producing being able to produce a "target" string for the found
element.

That is, the current point of access for resolving an "AutoLink",
`XmlGenerator::getAutoLink`, seems to be able to recognize an element
that the "AutoLink" candidate refers to, without necessarily producing a
link to the found target.

The states and reasons for which this happens are currently unknown and
should be addressed separately.

When this is the case, `AddAtomElements` would try to produce a link,
fail to do so and not fallback to writing the bare candidate.
In the more general case where a link was not produced because no
element that could tie to the candidate was found, the correct fallback
was generally chosen.

To avoid the issue, it is now ensured that, indepedently of there
being a link or not, we always write the raw word for the candidate,
either as the link text or as an unlinked raw string.

The scope of some variables that were used during the processing of an
"AutoLink" candidate in `addAtomElements` was slightly expanded to allow
the flattening of some of the branches around them.

Fixes: QTBUG-105278
Pick-to: 6.4
Change-Id: I265095cf42bf530878a6f741043ccbc7ee4cbb09
Reviewed-by: Paul Wicking <paul.wicking@qt.io>