QDoc/DocBook: Deal with QML style properties/methods
Don't output an anchor if the bridgehead has the same ID.
A similar fix was already implemented for sections.

This transformation cannot be implemented in macros as long as
CONFIG_DOCBOOKEXTENSIONS exists, because it would not be possible
to decide (within the configuration file) whether the synopsis should
be generated. This is not a real problem, as DocBook 5.2 will soon be
out (2022, depending on OASIS ratification process).

Change-Id: Iec41651d0715de8fc6cab576ab218d69bf6862a4
Reviewed-by: Paul Wicking <paul.wicking@qt.io>
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>