QDoc: Output sensible DocBook for raw strings
- CSS declarations.
- Exponents.
- Don't escape RawString that are not recognized:
  they might correspond to DocBook variants of custom macros.

Having dedicated code for this transformation helps qdoc work with
outdated clones of Qt repos and previous versions of Qt. This is an
important use case for me. Also, this could be important for other
qdoc users until the macro changes in many Qt Git repos are merged.
(The same comment remains valid for the other Atom::RawString changes.)

Change-Id: I8ef8520486390f5ba3d09508fe8e38355cf10b3f
Reviewed-by: Paul Wicking <paul.wicking@qt.io>