QDoc: Remove Node::operators
The Node class, the base class for QDoc's internal representation of
most documentation element, contains a static private map from string to
string called "operators".

The map was initialized once during the construction of the first Node
instance and seemed to contain a mapping between symbols representing
some C++'s operators and a "human readable naming of them".

The map doesn't see any use other than its initialization in the
codebase.

It is currently unclear what the purpose of the map was at its
introduction and when it might have lost its meaning.

Nonetheless, it is now removed as dead code.

The removal does not impact QDoc's compilation or the generation of the
current Qt's documentation, nor any regression test that is currently in place.

Task-number: QTBUG-105384
Change-Id: I44ccf95ebe9c7b719c5a3c65fb385bcf29200550
Reviewed-by: Paul Wicking <paul.wicking@qt.io>