QDoc: Move Node::setNavigationParent/navigationParent to PageNode
QDoc allows connecting the pages that it generates during an execution
to provide for an easier navigation between sequentially related pages.

In particular, it allows specifying, for each page, a page that is
"conceptually before" and a page that is "conceptually after",
through the use of the of the "\previouspage" and "\nextpage" commands.

Furthermore, QDoc provides, for the HTML format, an accessible "list" of
pages that can be traversed to arrive at the current viewed page, in the
form of a "breadcrumb list" or "navigation bar", to further enhance the
user ability to traverse backward when exploring the generated
documentation.

On the code side, this is done through multiple phases. The *page
commands are resolved during the parsing phase, as other "meta" commands
are, some internal state is set after the construction of the
`QDocDatabase` and some further processing is performed, based on the
previously set state, during the generation.

In particular, `Node` classes are able to store a single
"navigationParent", a pointer to another node representing a page that
is "conceptually before" the current node and that is expected to be
included in the "navigation bar".

Albeit the "navigationParent" is currently storable for all Node
classes, it is implicitly assumed that it will contain a `PageNode`, a
`Node` that will generate an actual page in the output documentation, as
this form of navigation is specifically designed for page-level
top-navigation, while other systems can be used for different form of
navigation, such as how inline links are used for intra-page navigation
or inter-page, sometimes element, navigation.

Indeed, most of the code that deals with the "navigationParent" is hidden
behind `PageNode`s checks.

This is in general true for other "relation" constructs, such as
the *page commands, which are meaningful at the page level.

While it is not unthinkable that there may be a use in allowing a more
granular link-ability for a "navigation bar", we currently expect this
to not be the case, and expect that a slightly different system might be
designed if it so were.

Hence, as it has been recently done for similar patterns in the
codebase, `Node::setnavigationParent/navigationParent`, the
property-methods for the "navigationParent", are now moved to the smaller
scope of a `PageNode`, where it is meaningful, removing it from the more
general `Node` class.

The only consumer of the "navigationParent",
`HtmlGenerator::generateNavigationBar`, was modified to cast the
currently processed `Node` to a `PageNode` explicitly, so as to be able
to keep using consuming the `navigationParent`.

The only producer of the "navigationParent",
`QDocDatabase::updateNavigation`, was modified to cast the processed
node to a `PageNode` to be able to access `setNavigationParent`.

Some of the condition for branching in `QDocDatabase::updateNavigation`
were simplified or moved around due to the ordering change that derives
from the explicit casting.

Similarly, the internal structures used by
`QDocDatabase::upadateNavigation` were modified to work with a
`PageNode` instead of a more general `Node`.

A branching condition to end the internal while loop in
`QDocDatabase::updateNavigation` was removed as duplicate code.
The condition would stop the loop when the last atom of the `Text`, as
per `Text::lastAtom` was reached by the loop.
The last atom of a `Text` is expected to point to `nullptr`, as no other
`Atom` will follow it, such that the loop will stop autonomously when
failing its condition which requires the currently iterated `Atom` to
not be a `nullptr`.

Some annotations comments were left in the code to mark possible further
work that was identified but is infeasible to evaluate at the current time.

Task-number: QTBUG-105384
Change-Id: I036bfa94ea1d20edd772df37606abebf68a913c5
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>