qtattributionsscanner: Allow */ to appear in license texts
License text files might contain */ (end of C style comment) , for
example when explaining how to attribute a license in a file. So far,
this broke the qdoc output, because qdoc also looks for */ as end of
documentation marker.

Work around this by masking the star through a qdoc \badcode argument,
so that

 */

becomes

 \badcode *
 \1/
 \endcode

Pick-to: 6.4
Change-Id: I77310ed272177eaa69e354c0d7bce380ee933e56
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Luca Di Sera <luca.disera@qt.io>