QDoc: Remove support for javascript codeblocks
QDoc provides an undocumented, not-officially-support, command pair,
"\js"/"\endjs", that allows a writer to include a codeblock of
javascript code in the documentation.

This is part of a suite of commands to allow pure javascript support
that was never finalized and is now being removed from the codebase.

Hence, all code related to the command pair is now removed and support
for it is dropped.

Some of the input data in the regression suite was modified as it made
use of a single instance of the command pair.
The regression files were regenerated to take into account the removal.

Fixes: QTBUG-106275
Change-Id: I1e18009617884a3841b0e2b83d4f5e34c7e14971
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>