QDoc: Remove Node::initialize
`Node`, the base class for documentable elements, has a static
`initialize` method, that is supposed to be called before using
instances of the `Node` class, whose purpose was to initialize static
members of the class.

In particular, the sole purpose was to initialize the "goal maps", a map
that is used by `LinkAtom`s, a class representing certain types of
to-resolve links found in a comment-block, to specify how to resolve
their containing link.

Supposedly, there was no valid constructor for a `QMap`, when this code
was added many years ago, to perform the same form of initialization in
place.

As this is not the case anymore, `Node::initialize` is now removed in
favor of initializing the static "goal map" in place.

`QDocDatabase::initializeDB`, that initialized, between other things,
`Node`, was modified to not call `Node::initialize` anymore, as a
consequence of the removal.

Change-Id: Iff87eeb2a9089bc8710e05269f0de8d069597609
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>