QDoc: Remove Node::goals
`Node`, the base class for documentable elements, provides a static
member, `goals`, that maps a "human-readable-string-representation" of a
`Node::NodeType` to a corresponding `Node::NodeType`, an enumeration
used to define the kind of element that the `Node` represents.

The "map of goals" had only a specific use, allowing a user of a "\l"
command to decided which "documentable element" the link should link to.

QDoc's link command, "\l", allows a user to create hypelinks between the
generated documentation pages in a project.
Optionally, the user can specify a series of "filters", that condition
the way in which QDoc finds the element that the link refers to.

For example, in a mixed C++ and QML project, there might a C++ type with
a certain name and a QML type with the same name, QDoc would prioritize
one of the two type when resolving a link specifying the name of that
element, but the user might require a link to the one that is not
prioritized.
By specifying the a C++ or QML domain, the user can ensure that the link
will be resolved to the correct one.

This is done by providing a bracketed argument to the "\l" command, with
a series of space delimited words that are used to resolve the filters.

`Node::goals` was used in this context. If the user provided a string,
in the bracketed arguments, that resolved to one of the keys in the map,
the resolving the link would be filtered to elements that have the
corresponding `Node::NodeType`.

Nonetheless, while the computation for the filter was performed, the
`Node::NodeType` filter was not actually used when resolving the link.

Furthermore, this particular syntax for the bracketed arguments of the
"\l" command was not even documented.

While it might be useful to allow the user to filter a link up to
specific element kind that they want to link to, a requirement for it
has never been present, up to this point, in Qt's documentation.

Hence, considering that the feature is unfinished, officially not
supported and not currently required or expected to be required in the
short-term, we remove `Node::goals` and the relevant surrounding code.

Thus, `Node::goals`, the static memmber of `Node` representing the
mapping between the "human-readable-string-representation" of a
`Node::NodeType` and its corresponding `Node::NodeType`, was removed.

The static member of `Node`, `goal`, which provided public access to
query `Node::goals` based on a string, was removed as a consequence of
the removal of `Node::goals`.

The code in `LinkAtom::resolveSquareBracketParams`, where the bracketed
parameters of a "\l" command were parsed and resolved, that resolved the
filter for a user-provided "Node::NodeType" string was removed to
disallow the hidden feature.

`LinkAtom::m_goal`, which would retain the `Node::NodeType` that should
be used as a filter when resolving the represented link, was removed as
it is now unnecessary.
As a consequence of the removal, the constructors of `LinkAtom` where
modified to not initialize `m_goal`.

Change-Id: I58fa3b953731b8185615a1cc4a868c78b8757e37
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>