QDoc: Remove deprecated usage of _qs literal operator
The usage was recently added but, Qt 6.8 onward, deprecates the usage of
"_qs" in favor of the "_s" literal operator in the
"Qt::Literals::StringLiterals" namespace.

A "using" for the namespaces was added to gain access to the
non-deprecated literal operators and the usage of "_qs" was converted to
"_s".

Change-Id: I43ad074ee7fdd1d50ab218cc74b57d983edca8da
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>