QDoc: Remove the "\newcode" and "\oldcode" command pair
The pair was deprecated in eb99338568dab0bd62144e1009974d5317b53037 and
was marked as to be removed in Qt 6.4.

As the release of 6.4 is near, we now remove all remaining code related
to the two commands.

[ChangeLog][qdoc] The deprecated \oldcode, \newcode commands were removed.

Fixes: QTBUG-98500
Change-Id: I06525d15836f07abca5ddfebe9eb557e9e9909a0
Reviewed-by: Kai Koehne <kai.koehne@qt.io>
(cherry picked from commit 356531bab8fd1c7ae53b0d5cd20459d5ce0a2db1)