QDoc: Devirtualize Node::getMemberNamespaces/getMemberClasses
The `Node` class, representing the base class in the hierarchy of
documentable elements, provides two virtual method,
`getMemberNamespaces` and `getMemberClasses`.

Both methods, are intended to be used only with `Node`s that have a
concept of containing multiple other `Node`s, where they allow the user
to find contained elements that have a specific characteristic, such as
being a "C++ Namespace".

In particular, they are intended to be used only by a `CollectionNode`,
a family of classes that represent entities that collect other `Node`s
into a single entity, such as the "groups" generated by "\group"
commands.

No other family of `Node`s than `CollectionNode` has a use for an
implementation of the two methods.

Consequently, we remove the virtual methods from the base `Node` class
and promote the concrete implementation of the methods in
`CollectionNode` to first-class, non-overridden members.

Change-Id: Ie87b42c2f7d0eac40b2eebc01485b4a8eda4b1b3
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>