QDoc: Remove unused since section
QDoc represents "sections", a part of the layout of a produced page, for
example the collection of documentations of the member of a class,
through the use of a class, `Section`, that contains a series of layout
information such as a title and is able to contain a series of elements
that compose it.

Furthermore, a class, `Sections`, is used as the main interface of
`Section`, acting as a container for all known "sections" that QDoc
uses, providing ways to access the data so that it can be converted
during the generation phase.

Of the various `Section`s that QDoc uses, there are some that are used
to layout a "\sincelist" command.

A "\sincelist" commands produces a list of elements that have been added
in a certain version of Qt.
This list is laid out in "sections", generally organized by element,
such that all classes appear under a certain heading, all Qml properties
appear under another and so on.

Internally, the section forming a "\sincelist" are represented as a
vector of `Section`s under the `Sections` class.

The `Section`s that for this vector are initialized so that they contain
a title for the sections, for example "New Type Aliases", and then
filled with the relevant elements for that section.

During the initialization, each `Section` of the vector is placed in a
certain order based on the integer representation of the elements of an
enumeration, `Sections::VectorIndex`.

The enumeration contains an element for each "since section" and all of
those elements are used to initialize one section of the vector of
"sections" for a "\sincelist".

Two elements of the enumeration, `VectorIndex::SinceTypedefs` and
`VectorIndex::SinceTypeAliases`, were sharing the same index.

Albeit two different initializations were provided for those two
members, since they are positioned based on their integer
representation, only one of the two initialization would actually be
represented as a section, as the latter initialization would overwrite
the original one.

In this case, the section containing "type-aliases" was initialized
later, meaning that it would be the one appearing in the generated
documents.

Due to the way in which sections are accessed and populated, which too
is based on the integer representation of elements of
`Sections::VectorIndex`, this did not generate any loss of
documentation, instead ending up showing both "typedefs" and
"type-aliases" under the same "section".

Albeit this did not produce any "visible error", the code is to still be
considered bugged, as it is not producing a section that is actually
represented in the code.

Instead of correcting the code to reintroduce the missing `Section`, we
remove references to the `Section` in the code, to avoid retroactively
modifying the output of the current documentation and as we consider the
current interface to be simpler for the user.

This preserves the current output that was used up until now, where both
"typedefs" and "type-aliases" are under the same "section" and removes the
"bugged" and confusing code.

To do so, `VectorIndex::SinceTypedefs` was removed and the two
references to it, to initialize the "since" `Section`s vector and to
dispatch elements to it afterwards, were removed as a consequence.

Change-Id: I868ba0f2e8d82422948b4e9c49fe1cb9764f7ff6
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>